// swapi/src/app/models/starships.ts
export interface IStarship {
  name: string;
  title: string;
  model: string;
  starship_class: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  crew: string;
  passengers: string;
  max_atmosphering_speed: string;
  MGLT: string;
  cargo_capacity: string;
  consumables: string;
  created: string;
  edited: string;
}